--- a/management/server/dns.go
+++ b/management/server/dns.go
@@ -4,11 +4,8 @@ import (
 	"context"
 	"fmt"
 	"strconv"
-	"strings"
 	"sync"
 
-	"github.com/hashicorp/go-multierror"
-	"github.com/miekg/dns"
 	log "github.com/sirupsen/logrus"
 
 	nbdns "github.com/netbirdio/netbird/dns"
@@ -19,30 +16,6 @@ import (
 )
 
 const defaultTTL = 300
-
-// CustomZoneCache is a thread-safe cache for custom DNS zones
-type CustomZoneCache struct {
-	cache sync.Map
-}
-
-// Get retrieves a cached custom zone
-func (c *CustomZoneCache) Get(key string) (nbdns.CustomZone, bool) {
-	if c == nil {
-		return nbdns.CustomZone{}, false
-	}
-	if value, ok := c.cache.Load(key); ok {
-		return value.(nbdns.CustomZone), true
-	}
-	return nbdns.CustomZone{}, false
-}
-
-// Set stores a custom zone in the cache
-func (c *CustomZoneCache) Set(key string, value nbdns.CustomZone) {
-	if c == nil {
-		return
-	}
-	c.cache.Store(key, value)
-}
 
 // DNSConfigCache is a thread-safe cache for DNS configuration components
 type DNSConfigCache struct {
@@ -251,52 +224,6 @@ func convertToProtoNameServerGroup(nsGroup *nbdns.NameServerGroup) *proto.NameSe
 		})
 	}
 	return protoGroup
-}
-
-func getPeersCustomZone(ctx context.Context, account *Account, dnsDomain string) nbdns.CustomZone {
-	var merr *multierror.Error
-
-	if dnsDomain == "" {
-		log.WithContext(ctx).Error("no dns domain is set, returning empty zone")
-		return nbdns.CustomZone{}
-	}
-
-	customZone := nbdns.CustomZone{
-		Domain:  dns.Fqdn(dnsDomain),
-		Records: make([]nbdns.SimpleRecord, 0, len(account.Peers)),
-	}
-
-	domainSuffix := "." + dnsDomain
-
-	var sb strings.Builder
-	for _, peer := range account.Peers {
-		if peer.DNSLabel == "" {
-			merr = multierror.Append(merr, fmt.Errorf("peer %s has an empty DNS label", peer.Name))
-			continue
-		}
-
-		sb.Grow(len(peer.DNSLabel) + len(domainSuffix))
-		sb.WriteString(peer.DNSLabel)
-		sb.WriteString(domainSuffix)
-
-		customZone.Records = append(customZone.Records, nbdns.SimpleRecord{
-			Name:  sb.String(),
-			Type:  int(dns.TypeA),
-			Class: nbdns.DefaultClass,
-			TTL:   defaultTTL,
-			RData: peer.IP.String(),
-		})
-
-		sb.Reset()
-	}
-
-	go func() {
-		if merr != nil {
-			log.WithContext(ctx).Errorf("error generating custom zone for account %s: %v", account.Id, merr)
-		}
-	}()
-
-	return customZone
 }
 
 func getPeerNSGroups(account *Account, peerID string) []*nbdns.NameServerGroup {