## Describe your changes

old:

```
$ go test -run '^$' -bench=BenchmarkUpdateAccountPeers ./management/server
goos: linux
goarch: amd64
pkg: github.com/netbirdio/netbird/management/server
cpu: 13th Gen Intel(R) Core(TM) i7-1355U
BenchmarkUpdateAccountPeers/Small-12                 568                 1.826 ms/op
BenchmarkUpdateAccountPeers/Medium-12                  9               139.3 ms/op
BenchmarkUpdateAccountPeers/Large-12                   1             15933 ms/op
BenchmarkUpdateAccountPeers/Small_single-12          198                 5.620 ms/op
BenchmarkUpdateAccountPeers/Medium_single-12           2               569.7 ms/op
BenchmarkUpdateAccountPeers/Large_5-12                 1             25845 ms/op
PASS
ok      github.com/netbirdio/netbird/management/server  59.603s
```

new:
```
$ go test -run '^$' -bench=BenchmarkUpdateAccountPeers ./management/server
goos: linux
goarch: amd64
pkg: github.com/netbirdio/netbird/management/server
cpu: 13th Gen Intel(R) Core(TM) i7-1355U
BenchmarkUpdateAccountPeers/Small-12                 649                 1.824 ms/op
BenchmarkUpdateAccountPeers/Medium-12                 30                53.51 ms/op
BenchmarkUpdateAccountPeers/Large-12                   1              4022 ms/op
BenchmarkUpdateAccountPeers/Small_single-12          664                 4.034 ms/op
BenchmarkUpdateAccountPeers/Medium_single-12           4               311.5 ms/op
BenchmarkUpdateAccountPeers/Large_5-12                 1              7261 ms/op
PASS
ok      github.com/netbirdio/netbird/management/server  28.074s
```

| Benchmark                         | Old Time (ms) | New Time (ms) | Improvement (%) |
|-----------------------------------|----------------|---------------|-----------------|
| Small                             | 1.826          | 1.824         | 0.11%           |
| Medium                            | 139.3          | 53.51         | 61.59%          |
| Large                             | 15933          | 4022          | 74.76%          |
| Small_single                      | 5.620          | 4.034         | 28.22%          |
| Medium_single                     | 569.7          | 311.5         | 45.32%          |
| Large_5                           | 25845          | 7261          | 71.90%          |

## Issue ticket number and link

### Checklist
- [ ] Is it a bug fix
- [ ] Is a typo/documentation fix
- [x] Is a feature enhancement
- [x] It is a refactor
- [ ] Created tests that fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary